Various changes for new statistics tables.
git-svn-id: https://svn.code.sf.net/p/ngsgenomelibs/code/trunk@594 2b8d95ea-eb58-0410-8a55-29fbd2c437c3

+++ perl_utils_usc/generateBamsToMethylChromTables.pl
@@ -5,21 +5,26 @@ use File::Temp qw/ tempfile tempdir /;
 use strict;
 use Getopt::Long;
 
-my $USAGE = "generateBamsToMethylChromTables.pl [--noremoveTemps] [--useReadCGschema] [--dumpCoverageMatrices] tablePrefix f1.bam f2.bam f3.bam ...";
+my $USAGE = "generateBamsToMethylChromTables.pl [--addMdTags] [--noremoveTemps] [--useReadCGschema] [--dumpCoverageMatrices] tablePrefix f1.bam f2.bam f3.bam ...";
 my @regions = ("chr1", "chr2", "chr3", "chr4", "chr5", "chr6", "chr7", "chr8", "chr9", 
 	       "chr10", "chr11", "chr12", "chr13", "chr14", "chr15", "chr16", "chr17", "chr18", 
 	       "chr19", "chr20", "chr21", "chr22", "chrX", "chrY", "chrM");
 #@regions = ("chr20","chr21");
+#@regions = ("chrM");
+
+
 
 my $SAMDIR = "/home/uec-00/shared/production/software/samtools";
+my $REFGENOME = "~/genomes/hg18_unmasked/hg18_unmasked.plusContam.fa";
 my $RMTMPS = 0;
 my $useReadCGschema = 0;
 my $minCphFrac = 1.01;
 my $minCphCoverage = 10;
 my $dumpCoverageMatrices = 0;
-GetOptions ('removeTemps!' => \$RMTMPS, 'useReadCGschema!' => \$useReadCGschema, 'dumpCoverageMatrices!' => \$dumpCoverageMatrices, 'minCphFrac' => \$minCphFrac) || die "$USAGE\n";
+my $addMdTags = 0;
+GetOptions ('addMdTags!' => \$addMdTags, 'removeTemps!' => \$RMTMPS, 'useReadCGschema!' => \$useReadCGschema, 'dumpCoverageMatrices!' => \$dumpCoverageMatrices, 'minCphFrac' => \$minCphFrac) || die "$USAGE\n";
 
-print STDERR "removeTemps=${RMTMPS}\n";
+print STDERR "removeTemps=${RMTMPS}\naddMdTags=${addMdTags}\n";
 
 # Input params
 die "$USAGE\n" unless (@ARGV>=2);
@@ -40,6 +45,10 @@ print STDERR "HOlding all jobs on job $holdJobId\n";
 
 foreach my $region (@regions)
 {
+#     my $regionSleepSecs = 1;
+#     print STDERR "Starting chrom ${region}. Waiting $regionSleepSecs to let PBS catch its breath\n";
+#     sleep($regionSleepSecs);
+
     my $tmpDir = "./tmp_${newname}_${region}";
     `mkdir $tmpDir`;
     my @dependJobs = ($holdJobId);
@@ -47,7 +56,7 @@ foreach my $region (@regions)
 
     # Generate regional bams within region directory
     my @regionalBamFns = ();
-#    my @dependJobsNext = ();
+    my @regionCmds = ();
     foreach my $bamFn (@bamFns)
     {
 	if ($bamFn !~ /\.bam$/)
@@ -57,12 +66,17 @@ foreach my $region (@regions)
 
 	my ($name, $path, $suf) = fileparse($bamFn, qr/\.bam/);
 	my $regionalBamFn = "${tmpDir}/${name}_${region}.bam";
-	my $pullRegionJobid = pullRegions($bamFn, $regionalBamFn, $region, $tmpDir,\@dependJobs);
-#	push(@dependJobsNext, $pullRegionJobid); # If you want to run them all at once
-	push(@dependJobs, $pullRegionJobid); # If you want to run all for the same chromosome in series (to not kill the disk)
+
+# Doing it all at once created more than a thousand jobs, which killed PBS       
+#	my $pullRegionJobid = pullRegions($bamFn, $regionalBamFn, $region, $tmpDir,\@dependJobs);
+#	push(@dependJobs, $pullRegionJobid); # If you want to run all for the same chromosome in series (to not kill the disk)
+
+	my $pullRegionCmd = pullRegionsCmdStr($bamFn, $regionalBamFn, $region, $tmpDir,\@dependJobs, $addMdTags);
+	push(@regionCmds,$pullRegionCmd);
 	push(@regionalBamFns, $regionalBamFn);
     }
-#    push(@dependJobs, @dependJobsNext);
+    my $pullRegionJobid = runCmd($tmpDir, join("; ",@regionCmds), "B2REG_pullRegionAndIndex", \@dependJobs);
+    push(@dependJobs, $pullRegionJobid);
 
     # Merge maps within region directory
     my $mergeName = "${tmpDir}/${newname}_${region}${processingSec}";
@@ -73,15 +87,15 @@ foreach my $region (@regions)
 
     if ($useReadCGschema)
     {
-    	$cmd .= "java -Xmx4000m edu.usc.epigenome.scripts.SamToMethylreadCGdbOffline -chrom ${region} -minMapQ 30 ${newname} $mergeBamFn";
+    	$cmd .= "java -Xmx3500m edu.usc.epigenome.scripts.SamToMethylreadCGdbOffline -chrom ${region} -minMapQ 30 ${newname} $mergeBamFn";
     }
     elsif ($dumpCoverageMatrices)
     {
-	$cmd .= "java -Xmx4000m edu.usc.epigenome.scripts.SamToConversionByCoverageMatrix -chrom ${region} -minMapQ 30 -outputCphs -minNextBaseCoverage 5 -minOppStrandCoverage 5 ${newname} $mergeBamFn";
+	$cmd .= "java -Xmx3500m edu.usc.epigenome.scripts.SamToConversionByCoverageMatrix -chrom ${region} -minMapQ 30 -outputCphs -minNextBaseCoverage 5 -minOppStrandCoverage 5 -maxCoverageOutput 300 ${newname} $mergeBamFn";
     }
     else
     {
-	$cmd .= "java -Xmx4000m edu.usc.epigenome.scripts.SamToMethyldbOffline -minConv 1 -useCpgsToFilter -minMapQ 30 -chrom ${region} ${newname} $mergeBamFn";
+	$cmd .= "java -Xmx3500m edu.usc.epigenome.scripts.SamToMethyldbOffline -minConv 1 -useCpgsToFilter -minMapQ 30 -chrom ${region} ${newname} $mergeBamFn";
     }
 
     my $curJobids = [runCmd($tmpDir,$cmd, "B2REG_createMehylTable", [$mergeJobid])];
@@ -89,10 +103,10 @@ foreach my $region (@regions)
 }
 
 # Now we're ready to release the hounds
-print STDERR "About to start jobs by releasing $holdJobId after a 60 second break\n";
-sleep(60);
+my $regionSleepSecs = 10;
+print STDERR "About to start jobs by releasing $holdJobId after a $regionSleepSecs second break\n";
+sleep($regionSleepSecs);
 `qrls $holdJobId`;
-
 
 
 # Returns
@@ -117,12 +131,16 @@ sub mergeBams
 	$curJobids = [runCmd($tmpdir, $cmd, "B2REG_mergeBams", $dependJobs)];
     }
 
-    # We have to remove dups again
-    my $curIn = $curOut;
-    $curOut = "${prefix}.NODUPS.bam";
-    my $cmd = "${SAMDIR}/samtools rmdup -s ${curIn} ${curOut}";
-    $cmd .= "; rm -f ${curIn}" if ($RMTMPS);
-    $curJobids = [runCmd($tmpdir,$cmd, "B2REG_rmdups", $curJobids)];
+    if (0)
+    {
+	# We have to remove dups again
+	my $curIn = $curOut;
+	$curOut = "${prefix}.NODUPS.bam";
+	my $cmd = "${SAMDIR}/samtools rmdup -s ${curIn} ${curOut}";
+	$cmd .= "; rm -f ${curIn}" if ($RMTMPS);
+	$curJobids = [runCmd($tmpdir,$cmd, "B2REG_rmdups", $curJobids)];
+    }
+
 
     my $finalBamOut = $curOut;
 
@@ -143,12 +161,51 @@ sub mergeBams
 
 
 
-# Returns the final process ID
-sub pullRegions
-{
-    my ($inBamFn, $outBamFn, $region, $tmpdir, $holdJobsIds) = @_;
+# # Returns the final process ID
+# sub pullRegions
+# {
+#     my ($inBamFn, $outBamFn, $region, $tmpdir, $holdJobsIds, $addMdTags) = @_;
 
-    my $finalProcId = 0;
+#     my $finalProcId = 0;
+
+#     if ($inBamFn !~ /\.bam$/)
+#     {
+# 	print STDERR "pipeline only works with .bam input files!\n";
+# 	exit;
+#     }
+
+#     my $curIn = $inBamFn;
+#     my $curOut = $outBamFn;
+#     my $curJobids = $holdJobsIds;
+
+#     # Put two commands in one job to half the number of jobs needed
+
+#     my $cmd1 = "${SAMDIR}/samtools view -b -o ${curOut} ${curIn} ${region}";
+# #    $curJobids = [runCmd($tmpdir,$cmd1, "B2REG_pullRegion", $curJobids)];
+
+#     if ($addMdTags)
+#     {
+# 	$curIn = $curOut;
+# 	$curOut =~ s/\.bam$/\.calmd\.bam/g;
+# 	$cmd1 .= "; ${SAMDIR}/samtools calmd -b ${curIn} ${REFGENOME} > ${curOut}";
+# 	$cmd1 .= "; rm -f ${curIn}" if ($RMTMPS);
+#     }
+
+#     # Index bam
+#     $curIn = $curOut;
+#     $curOut = "${curOut}.bai";
+#     my $cmd2 = "${SAMDIR}/samtools index ${curIn} ${curOut}";
+
+#     my $cmd = "${cmd1}; ${cmd2}";
+#     $curJobids = [runCmd($tmpdir,$cmd, "B2REG_pullRegionAndIndex", $curJobids)];
+
+#     $finalProcId = @${curJobids}[0];
+#     return $finalProcId;
+# }
+
+sub pullRegionsCmdStr
+{
+    my ($inBamFn, $outBamFn, $region, $tmpdir, $holdJobsIds, $addMdTags) = @_;
 
     if ($inBamFn !~ /\.bam$/)
     {
@@ -158,20 +215,29 @@ sub pullRegions
 
     my $curIn = $inBamFn;
     my $curOut = $outBamFn;
-    my $curJobids = $holdJobsIds;
 
-    my $cmd = "${SAMDIR}/samtools view -b -o ${curOut} ${curIn} ${region}";
-    $curJobids = [runCmd($tmpdir,$cmd, "B2REG_pullRegion", $curJobids)];
+    # Put two commands in one job to half the number of jobs needed
+
+    my $cmd1 = "${SAMDIR}/samtools view -b -o ${curOut} ${curIn} ${region}";
+
+    if ($addMdTags)
+    {
+	$curIn = $curOut;
+	$curOut =~ s/\.bam$/\.calmd\.bam/g;
+	$cmd1 .= "; ${SAMDIR}/samtools calmd -b ${curIn} ${REFGENOME} > ${curOut}";
+	$cmd1 .= "; rm -f ${curIn}" if ($RMTMPS);
+    }
 
     # Index bam
     $curIn = $curOut;
     $curOut = "${curOut}.bai";
-    my $cmd = "${SAMDIR}/samtools index ${curIn} ${curOut}";
-    $curJobids = [runCmd($tmpdir,$cmd, "B2REG_pullRegionIndex", $curJobids)];
+    my $cmd2 = "${SAMDIR}/samtools index ${curIn} ${curOut}";
 
-    $finalProcId = @${curJobids}[0];
-    return $finalProcId;
+    my $cmd = "${cmd1}; ${cmd2}";
+
+    return $cmd;
 }
+
 
 # Returns jobid
 sub runCmd
